requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
apturl==0.5.2
asgiref==3.8.1
asn1crypto==1.4.0
async-timeout==4.0.1
attrs==21.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==5.0.4
djangorestframework==3.15.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.21.2
drf-spectacular==0.27.2
drf_spectacular_extensions==0.0.2
duplicity==0.8.21
fasteners==0.14.1
flake8==4.0.1
frozenlist==1.2.0
future==0.18.2
gpg==1.16.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
jeepney==0.7.1
Jinja2==3.0.3
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mccabe==0.6.1
mdurl==0.1.2
mechanize==0.4.7
monotonic==1.6
more-itertools==8.10.0
multidict==5.1.0
netifaces==0.11.0
nh3==0.2.17
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pkginfo==1.10.0
protobuf==3.12.4
proton-core==0.1.15
proton-keyring-linux==0.0.1
proton-keyring-linux-secretservice==0.0.1
proton-vpn-api-core==0.20.1
proton-vpn-connection==0.11.0
proton-vpn-gtk-app==4.1.0
proton-vpn-killswitch==0.2.0
proton-vpn-killswitch-network-manager==0.2.0
proton-vpn-logger==0.2.1
proton-vpn-network-manager==0.3.0
proton-vpn-network-manager-openvpn==0.0.4
proton-vpn-session==0.6.1
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.8.0
pycups==2.0.1
pyflakes==2.4.0
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-gnupg==0.4.8
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
readme_renderer==43.0
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.5.0
systemd-python==234
tabulate==0.8.9
texttable==1.6.4
twine==5.0.0
typing_extensions==4.11.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.6
webencodings==0.5.1
websocket-client==1.2.3
xdg==5
xkit==0.0.0
yarl==1.7.2
zipp==1.0.0
